Added test support for an initial vector pubs legend
generate_legend_01.java here has been changed to create it at
https://url/vector/#11/51/33
No support for roads yet (that is a file, not generated).
The legend must be within --bbox -27.57,34.5,40.17,71.64
(or whatever the coastline bbox is)
so that when merged into existing tiles it does appear.

Not to be merged as it stands.
This is in Ukraine; the raster legend is within Australia.